pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Mindtickle assignment for the interview"
authors = ["Sanket Patole <snkt.pat2@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
typer = "^0.9.0"
psycopg2 = "^2.9.3"
mysql-connector-python = "^8.3.0"
pandas = "^2.2.2"
boto3 = "^1.34.83"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-mock = "^3.6"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
main = "src.scripts.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = ["__init__.py"]
